Drop nodegit, use lambda-git instead
`nodegit` caused all sorts of issues, and it takes too long to compile whether locally or on CI.  Even worse, due to [a bug in `serverless-webpack`](https://github.com/serverless-heaven/serverless-webpack/issues/309) we have to compile it twice, which takes over 15 minutes on Travis.

This PR drops `nodegit` and replaces it with a regular `git` binary [compiled specifically for AWS Lambda](https://github.com/pimterry/lambda-git).

This may or may not work!